requirements.txt
numpy
Cython
torch==1.7.0+cu101 torchvision==0.8.1+cu101 torchaudio==0.7.0 -f https://download.pytorch.org/whl/torch_stable.htmlCUDA=cu101
torch-scatter -f https://pytorch-geometric.com/whl/torch-1.7.0+cu110.html
torch-sparse -f https://pytorch-geometric.com/whl/torch-1.7.0+cu110.html
torch-cluster -f https://pytorch-geometric.com/whl/torch-1.7.0+cu110.html
torch-spline-conv -f https://pytorch-geometric.com/whl/torch-1.7.0+cu110.html
torch-geometric
h5py~=2.8.0
scipy~=1.3.1
open3d==0.10.0.0
pyyaml~=5.3.1
geomloss
Ninja
transforms3d